# public/api-docs/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "CRM RESTful API Docs v1.0",
        "description": "Just about CRM RESTful API Documentation. Enjoyed `Created by H@$b1`",
        "contact": {
            "email": "justhasby@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://157.245.55.185:8088/api/v1"
        }
    ],
...
    "paths": {
...
        "/pelayanan": {
            "post": {
                "tags": [
                    "Pelayanan-Controller"
                ],
                "summary": "getAllPelayananPaginate",
                "operationId": "getAllPelayananPaginate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/getIndexPaginate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "get": {
                "tags": [
                    "Pelayanan-Controller"
                ],
                "summary": "getPelayananById",
                "operationId": "getPelayananById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/parameters/getById"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pelayanan-Controller"
                ],
                "summary": "updatePelayanan",
                "operationId": "updatePelayanan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updatePelayanan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
...
        "/pelayanan/{id}": {
            "post": {
                "tags": [
                    "Pelayanan-Controller"
                ],
                "summary": "deletePelayananById",
                "operationId": "deletePelayananById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "$ref": "#/components/parameters/getByIdOnPath"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
...
        "/instansi": {
            "post": {
                "tags": [
                    "Instansi-Controller"
                ],
                "summary": "getAllInstansiPaginate",
                "operationId": "getAllInstansiPaginate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/getIndexPaginate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "get": {
                "tags": [
                    "Instansi-Controller"
                ],
                "summary": "getInstansiById",
                "operationId": "getInstansiById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/parameters/getById"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "put": {
                "tags": [
                    "Instansi-Controller"
                ],
                "summary": "updateInstansi",
                "operationId": "updateInstansi",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Instansi"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
...
    },
    "components": {
        "responses": {
            "400": {
                "description": "The Bad Request."
            },
            "401": {
                "description": "Unauthorized."
            },
            "403": {
                "description": "Forbidden."
            },
            "404": {
                "description": "The specified resource was not found."
            },
            "200": {
                "description": "Success."
            },
            "500": {
                "description": "Inter Server Error."
            }
        },
        "parameters": {
...
            "getById": {
                "name": "id",
                "in": "query"
            },
            "getByIdOnPath": {
                "name": "id",
                "in": "query",
                "required": true
            },
...
        },
        "schemas": {
            "getIndexPaginate": {
                "type": "object",
                "required": [
                    "page",
                    "size",
                    "sortField",
                    "sortOrder"
                ],
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "size": {
                        "type": "integer",
                        "example": 5
                    },
                    "sortField": {
                        "type": "string",
                        "example": "created_at"
                    },
                    "sortOrder": {
                        "type": "string",
                        "example": "desc"
                    },
                    "filters": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Filter"
                        }
                    }
                },
                "xml": {
                    "name": "getIndexPaginate"
                }
            },
            "Filter": {
                "type": "object",
                "required": [
                    "key",
                    "operation",
                    "value"
                ],
                "properties": {
                    "key": {
                        "type": "string",
                        "example": null
                    },
                    "operation": {
                        "type": "string",
                        "example": null
                    },
                    "value": {
                        "type": "string",
                        "default": null,
                        "example": null
                    }
                },
                "xml": {
                    "name": "Filter"
                }
            },
...
            "Instansi": {
                "type": "object",
                "properties": {
                    "id_instansi": {
                        "type": "string",
                        "description": "Input Id atau Kode Instansi"
                    },
                    "nama_instansi": {
                        "type": "string",
                        "description": "Input Tema"
                    },
                    "alamat": {
                        "type": "string",
                        "description": "Input Alamat"
                    },
                    "telp": {
                        "type": "integer",
                        "description": "Input telp"
                    }
                },
                "xml": {
                    "name": "Instansi"
                }
            },
...
            "updatePelayanan": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Input id"
                    },
                    "status": {
                        "type": "string",
                        "description": "Input status",
                        "default": 0,
                        "example": 0
                    },
                    "id_tracker": {
                        "type": "integer",
                        "description": "Input id_tracker",
                        "example": 1
                    },
                    "updated_by": {
                        "type": "string",
                        "description": "Input updated_by"
                    }
                },
                "xml": {
                    "name": "updatePelayanan"
                }
            },
...
        },
        "securitySchemes": {
            "CRM_Auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "CRM_Auth": []
        }
    ]
}